+++ lineage/DownwardRecursion.py
@@ -15,7 +15,7 @@ def get_gammas(tHMMobj, MSD, betas):
 
     for num, lO in enumerate(tHMMobj.X):  # for each lineage in our Population
         lineage = lO.output_lineage
-        MSDn = np.clip(MSD[num], np.finfo(np.float).eps, np.inf)
+        MSDn = np.clip(MSD[num], np.finfo(float).eps, np.inf)
 
         for level in lO.output_list_of_gens[1:]:
             for cell in level:
@@ -26,7 +26,7 @@ def get_gammas(tHMMobj, MSD, betas):
                     ci = lineage.index(d)
 
                     coeffs = betas[num] / MSDn
-                    beta_parent = np.clip(T @ coeffs[ci, :], np.finfo(np.float).eps, np.inf)
+                    beta_parent = np.clip(T @ coeffs[ci, :], np.finfo(float).eps, np.inf)
                     gammas[num][ci, :] = coeffs[ci, :] * np.matmul(gam / beta_parent, T)
     for gamm in gammas:
         assert np.all(np.isfinite(gamm))
